// src/app/shared/models/user.model.ts
import {UserTypeEnum} from 'src/app/shared/enums/user-type.enum';
import {VerificationStatusEnum} from '@sharedModule/models/graph-ql.model';
import {CEPOnboardingStateEnum} from '@campaigns/_enums/CEP-onboarding-state.enum';

export class UserModel {
	id: string;
	cognitoId: string;
	email: string;
	givenname: string;
	familyname: string;
	middlename: string;
	nickname: string;
	username: string;
	birthdate: string;
	gender: string;
	locale: string;
	mobileNumber: string;
	mobileDialCode?: string;
	mobileCountryCode?: string;
	profilePictureUrl: string;
	timezoneOffsetInMins: number;
	timezoneInfo: string;
	timezoneName: string;
	fbUserAccessToken: string;
	expiresAt: string;
	userType: UserTypeEnum | string;
	csAdminTeam: string | null;
	fbUserId: string;
	personaSurvey?: string | null;
	fullname: string;
	emailVerificationStatus: VerificationStatusEnum = VerificationStatusEnum.NotVerified;
	typeformResponseId?: string | null;
	whatsappSubscriptionStatus?: string | null;
	loginMethod?: string | null;
	productDemoedAtDate?: string | null;
	joinedCSGroupAtDate?: string | null;
	monetizationWorkshopAttendedAtDate?: string | null;
	receiveNotifications?: boolean | null;
	receiveWANotifications?: string | null;
	receiveEmailNotifications?: string | null;
	hasAccessToReportGenerationFeature?: boolean | null;
	totalWhatsappConfirmationRequest?: number | null;
	modeOfCommunication?: string | null;
	modeOfCommunicationVerificationStatus?: string | null;
	accessTokenReActivatedAtUTC?: string | null;
	createdAtUTC: string;
	isAdminBioCompleted?: boolean | null;
	fcmTokens: string[] | null = [];
	isEligibleForConvosightEarningPlatform?: boolean | null;
	CEPOnboardingState?: CEPOnboardingStateEnum;
	isBetaUser?: boolean;
	adminBioContactEmail?: string | null;
	adminBioContactPhoneNumber?: string | null;
	sendAdminBioNotificationCampaign?: boolean | null;
	ownsGroupProfile?: boolean | null;
}